refactor: Move build-id layout code to elf.rs
